IE Table: Made IE Table Address as 64 bit
IE Table address can be 64 bit. So hence for ls-ch3,
using 2 scratch registers (13 and 14) to store address of IE
table. Though IE Table address can only be 32 bit
(SG Table only support 32 bit address) but since uboot
recognize only 64 bit NOR Address, hence addresses in scratch needs
to be 64 bit.
Also corrected the address of SCRATCH13

+++ common/include/global.h
@@ -179,7 +179,7 @@ struct g_data_t {
 	uint32_t srk_size;
 	uint32_t ie_table_size;
 	uint32_t ie_table_offset;
-	uint32_t ie_table_addr;
+	uint64_t ie_table_addr;
 	uint32_t sg_offset;
 	uint32_t sg_size;
 	uint32_t sg_addr;
